+++ previous-versions/banking/RELEASES.md
@@ -337,3 +337,13 @@ Based on:
 - [go v0.25.0] previous-versions/banking
 ### Releases
 - [Go v0.25.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/banking/v0.25.0 - previous-versions/banking
+
+## 2023-11-20 01:30:33
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.25.1] previous-versions/banking
+### Releases
+- [Go v0.25.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/banking/v0.25.1 - previous-versions/banking